refactor: move uv discovery logic from workspace.js to python_uv.js
Move discoverUvWorkspaceMembers and hasProjectMetadata to their provider
file so workspace.js only retains generic scaffolding.
